Add a `send_email_task` task and use the `SendEmailTask` class as a base class.
Since `celery == 4.0.0`, class based tasks aren't automatically registered (https://docs.celeryq.dev/en/4.0/whatsnew-4.0.html#the-task-base-class-no-longer-automatically-register-tasks) so it is recommended to use the `task` (or `shared_task`) decorators.
This commit, introduces a `send_email_task` task (using the `shared_task`) decorator and sets the base of the task to be `SendEmailTask` (now renamed to `SendEmailBaseTask`).